+++ packages/react-ui/src/app/routes/platform/users/index.tsx
@@ -4,19 +4,13 @@ import { DataTable } from "@/components/ui/data-table";
 import { DataTableColumnHeader } from "@/components/ui/data-table-column-header";
 import { Tooltip, TooltipContent, TooltipTrigger } from "@/components/ui/tooltip";
 import { INTERNAL_ERROR_TOAST, useToast } from "@/components/ui/use-toast";
-import { platformHooks } from "@/hooks/platform-hooks";
 import { platformUserApi } from "@/lib/platform-user-api";
 import { formatUtils } from "@/lib/utils";
 import { useMutation } from "@tanstack/react-query";
 import { CircleMinus, Pencil, RotateCcw, Trash } from "lucide-react";
 import { useState } from "react";
-import { PlatformRole, UpdateUserRequestBody, UserStatus } from "../../../../../../shared/src";
-import { useForm } from "react-hook-form";
-import { typeboxResolver } from "@hookform/resolvers/typebox";
-import { Dialog, DialogContent, DialogFooter, DialogHeader, DialogTitle, DialogTrigger } from "@/components/ui/dialog";
-import { Form, FormField, FormItem, FormMessage } from "@/components/ui/form";
-import { Label } from "@/components/ui/label";
-import { Select, SelectContent, SelectGroup, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
+import { UserStatus } from "@activepieces/shared";
+import { UpdateUserRoleDialog } from "./update-role-dialog";
 
 export default function UsersPage() {
   const [refreshCount, setRefreshCount] = useState(0);
@@ -192,109 +186,3 @@ export default function UsersPage() {
   </div>
 }
 
-const UpdateUserRoleDialog = ({ children, onUpdate, userId, role }: { children: React.ReactNode, onUpdate: (role: PlatformRole) => void, userId: string, role: PlatformRole }) => {
-  const [open, setOpen] = useState(false);
-  const form = useForm<{ role: PlatformRole }>({
-    defaultValues: {
-      role,
-    },
-    resolver: typeboxResolver(UpdateUserRequestBody),
-  });
-  const { toast } = useToast();
-  const { mutate, isPending } = useMutation({
-    mutationKey: ['update-user'],
-    mutationFn: async (request: { platformRole: PlatformRole }) => {
-      await platformUserApi.update(userId, request);
-      onUpdate(request.platformRole);
-    },
-    onSuccess: () => {
-      setOpen(false);
-    },
-    onError: () => {
-      toast(INTERNAL_ERROR_TOAST);
-      setOpen(false);
-    },
-  });
-
-  return (
-    <Dialog
-      open={open}
-      onOpenChange={(open) => {
-        form.reset()
-        setOpen(open)
-      }}
-    >
-      <DialogTrigger asChild>
-        {children}
-      </DialogTrigger>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>Update User Role</DialogTitle>
-        </DialogHeader>
-        <Form {...form}>
-          <form
-            className="grid space-y-4"
-            onSubmit={(e) => e.preventDefault()}
-          >
-            <FormField
-              name="role"
-              render={({ field }) => (
-                <FormItem className="grid space-y-2">
-                  <Label htmlFor="role">
-                    Role
-                  </Label>
-                  <Select
-                    defaultValue={field.value} onValueChange={field.onChange}
-                    required
-                  >
-                    <SelectTrigger className="w-full">
-                      <SelectValue />
-                    </SelectTrigger>
-                    <SelectContent>
-                      <SelectGroup>
-                        {Object.values(PlatformRole).map((role) => (
-                          <SelectItem value={role} key={role}>
-                            {role}
-                          </SelectItem>
-                        ))}
-                      </SelectGroup>
-                    </SelectContent>
-                  </Select>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            {form?.formState?.errors?.root?.serverError && (
-              <FormMessage>
-                {form.formState.errors.root.serverError.message}
-              </FormMessage>
-            )}
-          </form>
-        </Form>
-        <DialogFooter>
-          <Button
-            variant={'outline'}
-            onClick={(e) => {
-              e.stopPropagation();
-              e.preventDefault();
-              setOpen(false);
-            }}
-          >
-            Cancel
-          </Button>
-          <Button
-            disabled={isPending}
-            loading={isPending}
-            onClick={(e) => {
-              e.stopPropagation();
-              e.preventDefault();
-              mutate({ platformRole: form.getValues().role });
-            }}
-          >
-            Save
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  )
-}
