+++ packages/react-ui/src/hooks/project-hooks.ts
@@ -58,13 +58,14 @@ const updateProject = async (
 const setCurrentProject = async (
   queryClient: QueryClient,
   project: ProjectWithLimits,
+  shouldReload = true,
 ) => {
   const projectChanged = authenticationSession.getProjectId() !== project.id;
   if (projectChanged) {
     await authenticationSession.switchToSession(project.id);
   }
   queryClient.setQueryData(['current-project'], project);
-  if (projectChanged) {
+  if (projectChanged && shouldReload) {
     window.location.reload();
   }
 };
